make a separate function to coerce clickmode and hovermode
